preserve NT_FILE note converting md to core
breakpad produced Minidumps have all the information needed to preserve the
NT_FILE note that linux coredumps normally have, but are not currently
included when converting using minidump-2-core. This adds a step to coredump
generation to add the note to the coredump. This change allows for two
improvements when loading coredumps into gdb:

1. for PIE code (default compilation now for gcc) gdb is now able to properly
validate the PIE displacement and apply it to the debugging session, providing
better stack trace quality.
2. gdb can now pick up so symbol files automatically without relying on the
scripting normally recommended by google for breakpad where add-symbol-file
is used to manually load symbols.

Bug: https://crbug.com/google-breakpad/766
Change-Id: I8cb25246dce0ae3492eedd6d3a4efcf1783d414d
Reviewed-on: https://chromium-review.googlesource.com/c/breakpad/breakpad/+/5463435
Reviewed-by: Mike Frysinger <vapier@chromium.org>